Fix setOutputPower()
Now power is an integer between 0 and PWMRANGE.

+++ src/BrushlessServo.h
@@ -20,7 +20,7 @@ class BrushlessServo {
 	void attach(int p1, int p2, int p3); // Initialization
     void writeOffset(int offset);  // Move to a offset between 0 and PRECISION
     void write(float degree);  // Move to a angle in degree
-    void setOutputPower(int p);  // Set a power multiplier between 0 (min) and 1024 (max)
+    void setOutputPower(int p);  // Set a power multiplier between 0 (min) and PWMRANGE (max)
 	void setCycles(int n);  // Set how many sinusoide period are needed for a revolution
 
   private:
@@ -29,7 +29,7 @@ class BrushlessServo {
     int pin1;
     int pin2;
     int pin3;
-    float power = PWMRANGE;
+    int power = PWMRANGE;
 	int n_cycles = 8;
 };
